+++ sentry/src/test/java/io/sentry/JsonSerializerTest.kt
@@ -9,6 +9,7 @@ import com.nhaarman.mockitokotlin2.verify
 import com.nhaarman.mockitokotlin2.whenever
 import io.sentry.exception.SentryEnvelopeException
 import io.sentry.protocol.Device
+import io.sentry.protocol.Request
 import io.sentry.protocol.SdkVersion
 import io.sentry.protocol.SentryId
 import io.sentry.protocol.SentrySpan
@@ -844,6 +845,50 @@ class JsonSerializerTest {
         val deserialized = fixture.serializer.deserialize(StringReader(serialized), SentryEvent::class.java)
 
         assertNull(deserialized?.threads)
+    }
+
+    @Test
+    fun `Long can be serialized inside request data`() {
+        val request = Request()
+
+        data class LongContainer(val longValue: Long)
+
+        request.data = LongContainer(10)
+
+        val serialized = serializeToString(request)
+        val deserialized = fixture.serializer.deserialize(StringReader(serialized), Request::class.java)
+
+        val deserializedData = deserialized?.data as? Map<String, Any>
+        assertNotNull(deserializedData)
+        assertEquals(10, deserializedData["longValue"])
+    }
+
+    @Test
+    fun `Primitives can be serialized inside request data`() {
+        val request = Request()
+
+        request.data = JsonReflectionObjectSerializerTest.ClassWithPrimitiveFields(
+            17,
+            3,
+            'x',
+            9001,
+            0.9f,
+            0.99,
+            true
+        )
+
+        val serialized = serializeToString(request)
+        val deserialized = fixture.serializer.deserialize(StringReader(serialized), Request::class.java)
+
+        val deserializedData = deserialized?.data as? Map<String, Any>
+        assertNotNull(deserializedData)
+        assertEquals(17, deserializedData["byte"])
+        assertEquals(3, deserializedData["short"])
+        assertEquals("x", deserializedData["char"])
+        assertEquals(9001, deserializedData["integer"])
+        assertEquals(0.9, deserializedData["float"])
+        assertEquals(0.99, deserializedData["double"])
+        assertEquals(true, deserializedData["boolean"])
     }
 
     @Test